segunda clase

+++ MiterceraClase.java
@@ -8,11 +8,11 @@ public class MiterceraClase {
         // System.out.println("division 7 / 5 =" +(7f/5f));
         Scanner pepe = new Scanner(System.in);
         System.out.print("Ingrese un número: ");
-        float numero1 = pepe.nextInt();
+        float numero1 = pepe.nextFloat();
 
 
         System.out.print("Ingrese un segundo numero: ");
-        float numero2 = pepe.nextInt();
+        float numero2 = pepe.nextFloat();
 
         System.out.println("la suma " + numero1 + " + "+ numero2 + " = " +(numero1+numero2) );
         System.out.println("la resta " + numero1 + " - "+ numero2 + " = " +(numero1-numero2) );
